requirements.txt
beautifulsoup4==4.13.3
Flask==3.1.0
gnews==0.4.0
jieba==0.42.1
line_bot_sdk==3.15.0
matplotlib==3.10.1
mediapipe==0.10.21
mplfinance==0.12.10b0
mysql_connector_repackaged==0.3.1
numpy==2.2.3
openai==1.65.4
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
protobuf==6.30.0
pymongo==4.11.2
Requests==2.32.3
scikit_learn==1.6.1
ta==0.11.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
twstock==1.4.0
wordcloud==1.9.4
yfinance==0.2.54
